user가 해당 게시물을 삭제할 수 있는지 여부에 따라서 버튼 표시를 판단, doDelete 방식 수정

+++ src/main/webapp/WEB-INF/jsp/usr/article/detail.jsp
@@ -4,7 +4,7 @@
 <%@ include file="../common/head.jspf"%>
 
 
-<div>${checkId}</div>
+<div>${checkUrlId}</div>
 
 <section class="mt-8 text-lg px-4">
 	<div class="mx-auto">
@@ -41,7 +41,9 @@
 			<c:if test="${article.userCanModify }">
 				<a href="../article/modify?id=${article.id }">수정</a>
 			</c:if>
-			<a href="../article/doDelete?id=${article.id }">삭제</a>
+			<c:if test="${article.userCanDelete }">
+				<a onclick="if(confirm('정말 삭제하시겠습니까?') == false) return false;" href="../article/doDelete?id=${article.id }">삭제</a>
+			</c:if>
 		</div>
 	</div>
 </section>
